chore(ui): remove dead ELEMENT_CONNECTED export and fix ShellContext TSDoc
Remove unused ELEMENT_CONNECTED constant from use-web-socket.ts.
Update ShellContext remarks to reflect that url is no longer a context
param — WebSocket URL is derived from self.location.origin internally.
